Add file "opt_IO/include/My_Class_VTK_02.hpp"
TODO: please change the name because that "_02" at the end can't be seen...

The file "opt_IO/include/My_Class_VTK_02.hpp" is the right formula of the
previous file "opt_IO/include/My_Class_VTK.hpp" to be used with cython, be-
cause this one, if the object passed by reference to the C++ library is a
Python object, will pass it by value, so removing all the reference after
the constructor call. That is why we changed the type of "data", being now a
pointer instead of a reference.