Add file "python_wrapper/opt_IO_impl/libPABLO.a" Add file "python_wrapper/opt_IO_impl/libopt_IO.a" Add file "python_wrapper/opt_IO_impl/libopt_base.a" Add file "python_wrapper/opt_IO_impl/setup.py"
Because the new cython files created in this folder need the library
"libopt_IO.a" (and "libopt_base.a" because the previous is based on this
last), and the library "libPABLO.a" (we are including the ".pyx" files we
need, that in this case and for now "class_para_tree.pyx", see sha1
0958d66a1dbe3450b012611d1ebff8536a3f3b7a), we have to modify the setup file
previously used in the "python_wrapper" directory.

Here we add two other user_option: "opt-base-include-path" and
"opt-IO-include-path", plus the two new libraries.