Modify file "python_wrapper/multiple_grids/utilities.py"
The file "utilities.py" now has a different parameters' list for function
"rendering_multi_block_data".

What is changed is that now "data" input parameter is a list, and for each
element in the list will be created a "RenderView" and, if this list has a
size bigger than one, will be added some "CameraLinks" between the
"RenderViews" to obtain the same scale in Paraview.