Modify file "python_wrapper/multiple_grids/block_matrices/project/utilities.py"
Added function "chunk_list_ordered". This function divides a list in an
ordered way, different from the function "chunk_list".

If for example we have the list [0, 1, 2, 3, 4, 5, 6, 7, 8, 9], and
"n_grids" is equal to 3, the new function returns:
[[0, 1, 2, 3], [4, 5, 6], [7, 8, 9]]
while the previous one would has returned the following chunk division:
[[0, 3, 6, 9], [1, 4, 7], [2, 5, 8]].

We have introduced this new function because, with the new block matrix, we
need processes of the same grid clustered together, to better use
preconditioning later.