Modify file "python_wrapper/multiple_grids/block_matrices/project/Laplacian2D.py"
Modified function "init_diag_mat".

Value inserted for the neighbor octant will be 0 if the current octant is
covered by grids of the first level (yes we are talking about background
grid). This because its solution value will be evaluated using the
coefficients of the bilinear interpolation of the octants of the upper
grid.

Also, it has been added the size of the edge "h" of the octants of the
background grid because it is useful when inserting the bilinear
coefficients from processes of the upper grids into lines of background
processes, because they have to be divided by "h2" of the line where they
are being inserted.