Modify file "python_wrapper/multiple_grids/block_matrices/project/Laplacian2D.py" Modify file "python_wrapper/multiple_grids/block_matrices/project/main.py"
To use multiple grids on the higher level, for the moment we have to
create one only inter-communicator for the grids of this level, because of
the MPI data type chosen for the receiving data. So, having one only inter-
communicator, grids of the first level will hqve the same tipology of
receiving data; To re-use multiple inter-communicator for the grids of the
first level, giving them the capability to speak each other, we have to
find a common MPI data type for data received from backround and from
foregrounds.