Modify file "python_wrapper/multiple_grids/block_matrices/project/main.py" Modify file "python_wrapper/multiple_grids/block_matrices/project/Laplacian2D.py" Modify file "python_wrapper/multiple_grids/block_matrices/config/PABLO.ini"
In order to make a patch for the description written in the commit SHA1
2a399f6abb1173b5cd7c7f4f2800d3bdc86550be, we have found a common MPI data
type for data received from backround and from foregrounds.

But because this would add data changed for MPI communications, we have
decided to give the user the ability to choose in the config file if there
is or no the particle interactions, for which the common MPI data type
is needed. The key used is "ParticlesInteraction". In additional, if the
interaction between particles is chosen, then we have to change the number
of intercommunicators for the foreground grids.